feat(classes): rename Vertex class getters for adjacent vertices and incident edges
Graph libraries such as NetworkX use the abbreviation "adj" for "adjacent", which in some cases makes properties and methods more readable. The Vertex getters for adjacent vertices have been renamed to begin with "adj_" instead of "adjacent_". The getters for incident edges have been renamed to start with "incident_" to add semantic clarity at the expense of short property names.

Closes #12